Keep the array-UDT test types under SuiteSparse 8's 128-byte cap
SuiteSparse built without variable-length arrays (MSVC, so the Windows
builds) rejects a UDT larger than 128 bytes with GrB_INVALID_VALUE before
9.0. The Windows job that drew psg 8.0.2 failed two tests on it:
test_nested_subarrays_register_flat (FP32[3, 2, 7], 168 bytes) and
test_udt_array_udf_known_extents_never_match_by_rank (FP64[21], 168 bytes).
test_udt_array_udf_returning_operand_keeps_its_type registers 152- and
160-byte types as well; it is marked slow, and that job did not run slow
tests.

None of the three is about size, so keep their shapes and shrink the
element type (INT16 or FP32) rather than skip them on SuiteSparse < 9.

+++ graphblas/tests/test_dtype.py
@@ -242,8 +242,10 @@ def test_nested_subarrays_register_flat():
     if _has_jit_set:
         assert udt.jit_c_definition is not None
 
-    deeper = np.dtype((np.dtype((np.dtype((np.float32, (7,))), (2,))), (3,)))
-    assert lookup_dtype(deeper).np_type == np.dtype((np.float32, (3, 2, 7)))
+    # INT16 keeps this under 128 bytes, the most SuiteSparse < 9 accepts for a
+    # UDT on builds without variable-length arrays (Windows).
+    deeper = np.dtype((np.dtype((np.dtype((np.int16, (7,))), (2,))), (3,)))
+    assert lookup_dtype(deeper).np_type == np.dtype((np.int16, (3, 2, 7)))
 
 
 def test_nested_subarrays_register_flat_in_records():
